AI-Python/backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
numpy==1.24.0
tensorflow-cpu==2.15.0
matplotlib==3.5.2
scikit-learn==1.2.2
pandas==2.0.3
